pyproject.toml
[project]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = [
    {name = "Lenar Khaidarov",email = "haidarov_lenar@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "faker (>=40.1.2,<41.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=26.1.0,<27.0.0)"
]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "coverage (>=7.13.2,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
